requirements.txt
Django>=5.0,<6.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
drf-yasg
channels
channels-redis
pillow
django-filter
psycopg2-binary
gunicorn
whitenoise
